requirements.txt
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
Pillow==12.0.0
scikit_learn==1.3.2
seaborn==0.13.2
torch==2.4.1+cu118
torchvision==0.19.1+cu118
